requirements.txt
langchain
python-dotenv
pypdf
openai
chromadb
tiktoken
gradio
gradio_pdf
qdrant_client
hupper
cohere
